move kotlin test logic
